+++ aiobotocore/_types.py
@@ -1,0 +1,3 @@
+from aiobotocore.httpsession import AIOHTTPSession
+
+DEFAULT_HTTP_SESSION_CLS = AIOHTTPSession